+++ CHANGELOG.md
@@ -6,11 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
 ## [1.0.8] - 2024-03-21
-
 ### Changed
+- Dashboard now logs when a call has started to prevent duplicate calls
 - Updated queries to match HiX 6.3 codes
+
+### Fixed
+- Bug where switching to new date didn't reset pred_idx
+- Bug where None values in call_number occured
 
 ## [1.0.7] - 2024-03-11